nyayalipi_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0